Extract block announce validation/polling from `ChainSync`
